Занятие 11/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.5.14
fastapi==0.90.1
fonttools==4.48.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
kiwisolver==1.4.5
langchain==0.0.271
langsmith==0.0.92
marshmallow==3.20.2
matplotlib==3.8.2
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.9.0
numpy==1.26.2
openai==0.27.4
packaging==23.2
pillow==10.2.0
pydantic==1.10.14
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
setuptools==69.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.23.1
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
wheel==0.41.3
yarl==1.9.4
faiss-cpu==1.7.4